Replace usage of fingerprints with groupIDs.
It turns out fingerprints are not always sent by sentry, and sometimes they're
not always unique (sometimes Sentry sends out the string '{{ default }}' as the
fingerprint). The prod Sentry service is not including the fingerprint in the
payloads being sent to SQS, for example.

Group ID happens to be an issue identifier we can use in its place, and it's
more consistently sent by Sentry.